lemon: remove similiar package
